+++ service/purchasing/src/main/java/org/icd4/commerce/application/required/cart/CartRepositoryPort.java
@@ -21,7 +21,6 @@ import java.util.Optional;
  * @author Jooeun
  * @since 1.0
  */
-@Repository
 public interface CartRepositoryPort {
     
     /**
