+++ service/purchasing/src/test/java/org/icd4/commerce/application/provided/cart/AddItemToCartUseCaseTest.java
@@ -4,11 +4,12 @@ import org.icd4.commerce.application.provided.cart.command.AddItemToCartCommand;
 import org.icd4.commerce.application.provided.cart.exception.CartNotFoundException;
 import org.icd4.commerce.application.provided.cart.exception.InsufficientStockException;
 import org.icd4.commerce.application.provided.cart.usecase.AddItemToCartUseCase;
-import org.icd4.commerce.application.required.CartRepositoryPort;
+import org.icd4.commerce.application.required.cart.CartRepositoryPort;
+import org.icd4.commerce.application.required.common.InventoryChecker;
 import org.icd4.commerce.domain.cart.*;
 import org.icd4.commerce.domain.cart.exception.CartAlreadyConvertedException;
 import org.icd4.commerce.domain.cart.exception.CartItemLimitExceededException;
-import org.icd4.commerce.domain.cart.exception.InvalidQuantityException;
+import org.icd4.commerce.domain.common.ProductId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
