+++ service/purchasing/src/test/java/org/icd4/commerce/application/provided/cart/ClearCartUseCaseTest.java
@@ -1,5 +1,8 @@
 package org.icd4.commerce.application.provided.cart;
 
+import org.icd4.commerce.application.provided.cart.command.ClearCartCommand;
+import org.icd4.commerce.application.provided.cart.exception.CartNotFoundException;
+import org.icd4.commerce.application.provided.cart.usecase.ClearCartUseCase;
 import org.icd4.commerce.application.required.cart.CartRepositoryPort;
 import org.icd4.commerce.domain.cart.Cart;
 import org.icd4.commerce.domain.cart.CartId;
