+++ service/purchasing/src/test/java/org/icd4/commerce/application/provided/cart/RemoveItemFromCartUseCaseTest.java
@@ -3,10 +3,11 @@ package org.icd4.commerce.application.provided.cart;
 import org.icd4.commerce.application.provided.cart.command.RemoveItemFromCartCommand;
 import org.icd4.commerce.application.provided.cart.exception.CartNotFoundException;
 import org.icd4.commerce.application.provided.cart.usecase.RemoveItemFromCartUseCase;
-import org.icd4.commerce.application.required.CartRepositoryPort;
+import org.icd4.commerce.application.required.cart.CartRepositoryPort;
 import org.icd4.commerce.domain.cart.*;
 import org.icd4.commerce.domain.cart.exception.CartAlreadyConvertedException;
 import org.icd4.commerce.domain.cart.exception.InvalidCartStateException;
+import org.icd4.commerce.domain.common.ProductId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
